add type cast in src/gromacs/mdlib/nbnxn_cuda/nbnxn_cuda_kernel_utils.cuh to fix the bug encountered on rocm 3.0

+++ src/gromacs/mdlib/nbnxn_cuda/nbnxn_cuda_kernel_utils.cuh
@@ -437,7 +437,7 @@ float interpolate_coulomb_force_r(const cu_nbparam_t nbparam,
 
     float2 d01 = fetch_coulomb_force_r(nbparam, index);
 
-    return lerp(d01.x, d01.y, fraction);
+    return lerp((float)d01.x, (float)d01.y, fraction);
 }
 
 /*! Fetch C6 and C12 from the parameter table.
